requirements.txt
asteval==1.0.6
attrdict==2.0.1
elasticsearch==9.0.2
emoji==2.14.1
ftfy==6.3.1
lingua-language-detector==2.0.2
mosestokenizer==1.2.1
multipledispatch==1.0.0
nltk==3.9.1
numpy
parlai
parse==1.20.2
protobuf==6.31.1
PyQt5==5.15.2
pytest==8.4.1
PyYAML==6.0.2
regex==2024.11.6
retico_core==0.2.10
retico_googleasr==0.1.5
scipy
sortedcontainers==2.4.0
soundcard==0.4.4
soundfile==0.13.1
spacy==3.8.7
speechbrain==1.0.3
torch==2.6.0
tqdm
TTS==0.22.0